- first attempt at uploading the octree collision code. Currently this just uses the octree to detect collisions, but still calculates some of the collision overlaps more than once. This can be optimized - Introduced a system for communicating between "multiple threads". If this is used, we could allow something like the module loader to run in a background thread and then the menu wouldn't completely freeze. I have implemented this in more than one private version, but... - Broke out several functions into *.inl files so that they could be more easily marked as inline files. In practice, this makes for a faster executible... but we will have to see. - Optimized the calling of several functions that were bogging down the execution, including the detection of wall collisions. - Introduced a method of measuring the total amount to time that each script takes, as well as the total time for each script function. This may help with determining if one or more of these is bogging down the program. - Removed some memory leaks from the ttf text display. - Formatted the code - Maybe some other things, but I can't remember.

+++ game/egoboo_console.inl
@@ -55,10 +55,10 @@ struct s_egoboo_console
     int    save_index;
     char   save_buffer[EGOBOO_CONSOLE_LINES][EGOBOO_CONSOLE_LENGTH];
 
-    int    buffer_carat;
+    size_t buffer_carat;
     char   buffer[EGOBOO_CONSOLE_LENGTH];
 
-    int    output_carat;
+    size_t output_carat;
     char   output_buffer[EGOBOO_CONSOLE_OUTPUT];
 };
